requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colors.py==0.2.2
idna==3.7
pygame==2.5.2
requests==2.31.0
urllib3==2.2.1